+++ SpaceShared/PositionAnchor.cs
@@ -1,0 +1,18 @@
+namespace SpaceShared
+{
+    /// <summary>The screen area from which a position is relative.</summary>
+    internal enum PositionAnchor
+    {
+        /// <summary>The bottom-left corner of the screen.</summary>
+        BottomLeft,
+
+        /// <summary>The bottom-right corner of the screen.</summary>
+        BottomRight,
+
+        /// <summary>The top-left corner of the screen.</summary>
+        TopLeft,
+
+        /// <summary>The top-right corner of the screen.</summary>
+        TopRight
+    }
+}